bestfriend still with flaws but algorithm working

+++ push_swap/push_swap.h
@@ -54,7 +54,7 @@ void            print_integer(int *content);
 void            insertion_sort(t_node **head);
 float           mean_of_nodes(t_node **head);
 void            ten_or_more(t_node **head_a, t_node **head_b);
-int             get_best_friend(t_node **head_a, long number);
+int             get_best_friend(t_node **head_a, int number);
 int	            get_index(t_node **head, int num);
 void            find_best_move(t_node **head_a, t_node **head_b);
 unsigned int    absolute(int number);
